Handle object values
The original `Signal` class worked perfectly for primitive types but the
simple `value` setter did not pick up when nested object properties were
updated or if array members were updated.
It only worked for full assignments / overwrites.

Split the `Signal` calss into a `PrimitiveSignal` and an `ObjectSignal`.
`PrimitiveSignal` remains unchanged from the original.
`ObjectSignal` uses the `Proxy` class to listen for updates to
deep properties and emit these changes to all connected slots.

Add a basic `createSignal` function as the main API that handles creating,
and returning an instance of the correct class based on the type of the
provided value.

Add more unit tests.